refactor(elixir): use top level address, message and route
