+++ implementations/elixir/ockam/ockam/lib/ockam/router.ex
@@ -7,12 +7,12 @@ defmodule Ockam.Router do
   for which there is no address type specific handler.
   """
 
-  import Ockam.Router.Guards
+  import Ockam.Address, only: [is_address_type: 1]
+  import Ockam.Router.MessageHandler, only: [is_message_handler: 1]
 
-  alias Ockam.Router.Address
-  alias Ockam.Router.Message
+  alias Ockam.Address
+  alias Ockam.Message
   alias Ockam.Router.MessageHandler
-  alias Ockam.Router.Route
   alias Ockam.Router.Storage
 
   alias Ockam.Telemetry
@@ -35,8 +35,8 @@ defmodule Ockam.Router do
   end
 
   defp pick_and_invoke_handler(message) do
-    first_address_type = message |> Message.onward_route() |> Route.first_address_type()
-    handler_type = if first_address_type, do: first_address_type, else: :default
+    first_address = message |> Message.onward_route() |> List.first()
+    handler_type = if first_address, do: Address.type(first_address), else: :default
 
     case get_message_handler(handler_type) do
       nil -> {:error, {:handler_not_set, handler_type, message}}
@@ -63,7 +63,7 @@ defmodule Ockam.Router do
 
   def get_message_handler(:default), do: Storage.get(:default_message_handler)
 
-  def get_message_handler(address_type) when is_address_type(address_type) do
+  def get_message_handler(address_type) when Address.is_address_type(address_type) do
     case Storage.get({:address_type_message_handler, address_type}) do
       nil -> get_message_handler(:default)
       handler -> handler
